+++ NEWS.md
@@ -4,7 +4,10 @@
 
 *12/??/2-2017*
 
-* :construction_worker:Updated tests to use asynbolt.Server. Test servers now close gracefully w/o error logs.:construction_worker:
+* :pencil2:Renamed `pack_message` and `unpack_message` to `serialize_message` and `deserialize_message:pencil2:
+* :recycle:Added a parser that handles message transfer encoding -> simplified read buffer logic:recycle:
+* :bug:Fixed bug in write buffer that allowed zero length chunks:bug: 
+* :construction_worker:Updated tests to use asyncbolt.Server. Test servers now close gracefully w/o error logs.:construction_worker:
 
 ## 0.0.1